Updated mktdata gateway and coinbase service #85
